Update SwiftyJSON to 2.2.0
Update SwiftyJSON to 2.2.0 to try to fix a Carthage building issue for
SwiftyJSON
