+++ Cartfile.resolved
@@ -1,2 +1,2 @@
-github "Alamofire/Alamofire" "1.2.1"
+github "Alamofire/Alamofire" "1.2.3"
 github "SwiftyJSON/SwiftyJSON" "2.2.0"
