+++ Carthage/Checkouts/Alamofire/.gitignore
@@ -1,5 +1,7 @@
-# Xcode
+# Mac OS X
+.DS_Store
 
+# Xcode
 build/
 *.pbxuser
 !default.pbxuser
@@ -16,3 +18,6 @@ DerivedData
 *.hmap
 *.ipa
 *.xcuserstate
+
+# Carthage
+Carthage/Build
