+++ Carthage/Checkouts/Alamofire/LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2014 Alamofire (http://alamofire.org/)
+Copyright (c) 2014–2015 Alamofire Software Foundation (http://alamofire.org/)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
