+++ Carthage/Checkouts/Alamofire/Source/Alamofire.h
@@ -1,6 +1,6 @@
 // Alamofire.h
 //
-// Copyright (c) 2014–2015 Alamofire (http://alamofire.org)
+// Copyright (c) 2014–2015 Alamofire Software Foundation (http://alamofire.org/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
