+++ Carthage/Checkouts/Alamofire/Tests/BaseTestCase.swift
@@ -24,25 +24,6 @@ import Alamofire
 import Foundation
 import XCTest
 
-class TLSEvaluationTestCase: BaseTestCase {
-    func testThatExpiredSSLCertificateFailsEvaluation() {
-        // Given
-        let URL = "https://testssl-expire.disig.sk/"
-        let expectation = expectationWithDescription("\(URL)")
-
-        var error: NSError?
-
-        // When
-        Alamofire.request(.GET, URL)
-            .response { _, _, _, responseError in
-                error = responseError
-                expectation.fulfill()
-        }
-
-        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
-
-        // Then
-        XCTAssertNotNil(error, "error should not be nil")
-        XCTAssertEqual(error?.code ?? -1, NSURLErrorServerCertificateUntrusted, "error should be NSURLErrorServerCertificateUntrusted")
-    }
+class BaseTestCase: XCTestCase {
+    let defaultTimeout: NSTimeInterval = 10
 }
