+++ Carthage/Checkouts/Alamofire/Tests/ManagerTests.swift
@@ -1,6 +1,6 @@
 // RequestTests.swift
 //
-// Copyright (c) 2014–2015 Alamofire (http://alamofire.org)
+// Copyright (c) 2014–2015 Alamofire Software Foundation (http://alamofire.org/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,12 +20,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Foundation
 import Alamofire
+import Foundation
 import XCTest
 
-class AlamofireManagerTestCase: XCTestCase {
+class ManagerTestCase: BaseTestCase {
     func testSetStartRequestsImmediatelyToFalseAndResumeRequest() {
+        // Given
         let manager = Alamofire.Manager()
         manager.startRequestsImmediately = false
 
@@ -34,27 +35,56 @@ class AlamofireManagerTestCase: XCTestCase {
 
         let expectation = expectationWithDescription("\(URL)")
 
+        var response: NSHTTPURLResponse?
+
+        // When
         manager.request(URLRequest)
-            .response { (_,_,_,_) in expectation.fulfill() }
+            .response { _, responseResponse, _, _ in
+                response = responseResponse
+                expectation.fulfill()
+            }
             .resume()
 
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
-        }
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertTrue(response?.statusCode == 200, "response status code should be 200")
     }
 
     func testReleasingManagerWithPendingRequestDeinitializesSuccessfully() {
+        // Given
+        var manager: Manager? = Alamofire.Manager()
+        manager?.startRequestsImmediately = false
+
+        let URL = NSURL(string: "http://httpbin.org/get")!
+        let URLRequest = NSURLRequest(URL: URL)
+
+        // When
+        let request = manager?.request(URLRequest)
+        manager = nil
+
+        // Then
+        XCTAssertTrue(request?.task.state == .Suspended, "request task state should be '.Suspended'")
+        XCTAssertNil(manager, "manager should be nil")
+    }
+
+    func testReleasingManagerWithPendingCanceledRequestDeinitializesSuccessfully() {
+        // Given
         var manager: Manager? = Alamofire.Manager()
         manager!.startRequestsImmediately = false
 
         let URL = NSURL(string: "http://httpbin.org/get")!
         let URLRequest = NSURLRequest(URL: URL)
 
+        // When
         let request = manager!.request(URLRequest)
-
+        request.cancel()
         manager = nil
 
-        XCTAssert(request.task.state == .Suspended)
-        XCTAssertNil(manager)
+        // Then
+        let state = request.task.state
+        XCTAssertTrue(state == .Canceling || state == .Completed, "state should be .Canceling or .Completed")
+        XCTAssertNil(manager, "manager should be nil")
     }
 }
