+++ Carthage/Checkouts/Alamofire/Tests/RequestTests.swift
@@ -20,121 +20,241 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Foundation
 import Alamofire
+import Foundation
 import XCTest
 
-class AlamofireRequestInitializationTestCase: XCTestCase {
+class RequestInitializationTestCase: BaseTestCase {
     func testRequestClassMethodWithMethodAndURL() {
-        let URL = "http://httpbin.org/"
-        let request = Alamofire.request(.GET, URL)
+        // Given
+        let URLString = "http://httpbin.org/"
 
+        // When
+        let request = Alamofire.request(.GET, URLString)
+
+        // Then
         XCTAssertNotNil(request.request, "request should not be nil")
-        XCTAssertEqual(request.request.URL!, NSURL(string: URL)!, "request URL should be equal")
+        XCTAssertEqual(request.request.URL!, NSURL(string: URLString)!, "request URL should be equal")
         XCTAssertNil(request.response, "response should be nil")
     }
 
     func testRequestClassMethodWithMethodAndURLAndParameters() {
-        let URL = "http://httpbin.org/get"
-        let request = Alamofire.request(.GET, URL, parameters: ["foo": "bar"])
+        // Given
+        let URLString = "http://httpbin.org/get"
 
+        // When
+        let request = Alamofire.request(.GET, URLString, parameters: ["foo": "bar"])
+
+        // Then
         XCTAssertNotNil(request.request, "request should not be nil")
-        XCTAssertNotEqual(request.request.URL!, NSURL(string: URL)!, "request URL should be equal")
-        XCTAssertEqual(request.request.URL!.query!, "foo=bar", "query is incorrect")
+        XCTAssertNotEqual(request.request.URL!, NSURL(string: URLString)!, "request URL should be equal")
+        XCTAssertEqual(request.request.URL?.query ?? "", "foo=bar", "query is incorrect")
         XCTAssertNil(request.response, "response should be nil")
     }
 }
 
-class AlamofireRequestResponseTestCase: XCTestCase {
+// MARK: -
+
+class RequestResponseTestCase: BaseTestCase {
     func testRequestResponse() {
-        let URL = "http://httpbin.org/get"
+        // Given
+        let URLString = "http://httpbin.org/get"
         let serializer = Alamofire.Request.stringResponseSerializer(encoding: NSUTF8StringEncoding)
 
-        let expectation = expectationWithDescription("\(URL)")
+        let expectation = expectationWithDescription("GET request should succeed: \(URLString)")
 
-        Alamofire.request(.GET, URL, parameters: ["foo": "bar"])
-                 .response(serializer: serializer){ (request, response, string, error) in
-                    XCTAssertNotNil(request, "request should not be nil")
-                    XCTAssertNotNil(response, "response should not be nil")
-                    XCTAssertNotNil(string, "string should not be nil")
-                    XCTAssertNil(error, "error should be nil")
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var string: AnyObject?
+        var error: NSError?
 
-                    expectation.fulfill()
+        // When
+        Alamofire.request(.GET, URLString, parameters: ["foo": "bar"])
+            .response(serializer: serializer) { responseRequest, responseResponse, responseString, responseError in
+                request = responseRequest
+                response = responseResponse
+                string = responseString
+                error = responseError
+
+                expectation.fulfill()
         }
 
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
-        }
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(string, "string should not be nil")
+        XCTAssertNil(error, "error should be nil")
     }
-}
 
-class AlamofireRequestDescriptionTestCase: XCTestCase {
-    func testRequestDescription() {
-        let URL = "http://httpbin.org/get"
-        let request = Alamofire.request(.GET, URL)
+    func testRequestResponseWithProgress() {
+        // Given
+        let randomBytes = 4 * 1024 * 1024
+        let URLString = "http://httpbin.org/bytes/\(randomBytes)"
 
-        XCTAssertEqual(request.description, "GET http://httpbin.org/get", "incorrect request description")
+        let expectation = expectationWithDescription("Bytes download progress should be reported: \(URLString)")
 
-        let expectation = expectationWithDescription("\(URL)")
+        var byteValues: [(bytes: Int64, totalBytes: Int64, totalBytesExpected: Int64)] = []
+        var progressValues: [(completedUnitCount: Int64, totalUnitCount: Int64)] = []
+        var responseRequest: NSURLRequest?
+        var responseResponse: NSHTTPURLResponse?
+        var responseData: AnyObject?
+        var responseError: NSError?
 
-        request.response { (_, response,_,_) in
-            XCTAssertEqual(request.description, "GET http://httpbin.org/get (\(response!.statusCode))", "incorrect request description")
+        // When
+        let request = Alamofire.request(.GET, URLString)
+        request.progress { bytesRead, totalBytesRead, totalBytesExpectedToRead in
+            let bytes = (bytes: bytesRead, totalBytes: totalBytesRead, totalBytesExpected: totalBytesExpectedToRead)
+            byteValues.append(bytes)
+
+            let progress = (completedUnitCount: request.progress.completedUnitCount, totalUnitCount: request.progress.totalUnitCount)
+            progressValues.append(progress)
+        }
+        request.response { request, response, data, error in
+            responseRequest = request
+            responseResponse = response
+            responseData = data
+            responseError = error
 
             expectation.fulfill()
         }
 
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(responseRequest, "response request should not be nil")
+        XCTAssertNotNil(responseResponse, "response response should not be nil")
+        XCTAssertNotNil(responseData, "response data should not be nil")
+        XCTAssertNil(responseError, "response error should be nil")
+
+        XCTAssertEqual(byteValues.count, progressValues.count, "byteValues count should equal progressValues count")
+
+        if byteValues.count == progressValues.count {
+            for index in 0..<byteValues.count {
+                let byteValue = byteValues[index]
+                let progressValue = progressValues[index]
+
+                XCTAssertGreaterThan(byteValue.bytes, 0, "reported bytes should always be greater than 0")
+                XCTAssertEqual(byteValue.totalBytes, progressValue.completedUnitCount, "total bytes should be equal to completed unit count")
+                XCTAssertEqual(byteValue.totalBytesExpected, progressValue.totalUnitCount, "total bytes expected should be equal to total unit count")
+            }
+        }
+
+        if let lastByteValue = byteValues.last,
+            lastProgressValue = progressValues.last
+        {
+            let byteValueFractionalCompletion = Double(lastByteValue.totalBytes) / Double(lastByteValue.totalBytesExpected)
+            let progressValueFractionalCompletion = Double(lastProgressValue.0) / Double(lastProgressValue.1)
+
+            XCTAssertEqual(byteValueFractionalCompletion, 1.0, "byte value fractional completion should equal 1.0")
+            XCTAssertEqual(progressValueFractionalCompletion, 1.0, "progress value fractional completion should equal 1.0")
+        } else {
+            XCTFail("last item in bytesValues and progressValues should not be nil")
         }
     }
 }
 
-class AlamofireRequestDebugDescriptionTestCase: XCTestCase {
+// MARK: -
+
+class RequestDescriptionTestCase: BaseTestCase {
+    func testRequestDescription() {
+        // Given
+        let URLString = "http://httpbin.org/get"
+        let request = Alamofire.request(.GET, URLString)
+        let initialRequestDescription = request.description
+
+        let expectation = expectationWithDescription("Request description should update: \(URLString)")
+
+        var finalRequestDescription: String?
+        var response: NSHTTPURLResponse?
+
+        // When
+        request.response { _, responseResponse, _, _ in
+            finalRequestDescription = request.description
+            response = responseResponse
+
+            expectation.fulfill()
+        }
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertEqual(initialRequestDescription, "GET http://httpbin.org/get", "incorrect request description")
+        XCTAssertEqual(finalRequestDescription ?? "", "GET http://httpbin.org/get (\(response?.statusCode ?? -1))", "incorrect request description")
+    }
+}
+
+// MARK: -
+
+class RequestDebugDescriptionTestCase: BaseTestCase {
+    // MARK: Properties
+
     let manager: Alamofire.Manager = {
         let manager = Alamofire.Manager(configuration: NSURLSessionConfiguration.defaultSessionConfiguration())
         manager.startRequestsImmediately = false
         return manager
     }()
 
-    // MARK: -
+    let managerDisallowingCookies: Alamofire.Manager = {
+        let configuration = NSURLSessionConfiguration.defaultSessionConfiguration()
+        configuration.HTTPShouldSetCookies = false
+
+        let manager = Alamofire.Manager(configuration: configuration)
+        manager.startRequestsImmediately = false
+
+        return manager
+    }()
+
+    // MARK: Tests
 
     func testGETRequestDebugDescription() {
-        let URL = "http://httpbin.org/get"
-        let request = manager.request(.GET, URL)
+        // Given
+        let URLString = "http://httpbin.org/get"
+
+        // When
+        let request = manager.request(.GET, URLString)
         let components = cURLCommandComponents(request)
 
-        XCTAssert(components[0..<3] == ["$", "curl", "-i"], "components should be equal")
-        XCTAssert(!contains(components, "-X"), "command should not contain explicit -X flag")
-        XCTAssert(components.last! == "\"\(URL)\"", "URL component should be equal")
+        // Then
+        XCTAssertEqual(components[0..<3], ["$", "curl", "-i"], "components should be equal")
+        XCTAssertFalse(contains(components, "-X"), "command should not contain explicit -X flag")
+        XCTAssertEqual(components.last ?? "", "\"\(URLString)\"", "URL component should be equal")
     }
 
     func testPOSTRequestDebugDescription() {
-        let URL = "http://httpbin.org/post"
-        let request = manager.request(.POST, URL)
+        // Given
+        let URLString = "http://httpbin.org/post"
+
+        // When
+        let request = manager.request(.POST, URLString)
         let components = cURLCommandComponents(request)
 
-        XCTAssert(components[0..<3] == ["$", "curl", "-i"], "components should be equal")
-        XCTAssert(components[3..<5] == ["-X", "POST"], "command should contain explicit -X flag")
-        XCTAssert(components.last! == "\"\(URL)\"", "URL component should be equal")
+        // Then
+        XCTAssertEqual(components[0..<3], ["$", "curl", "-i"], "components should be equal")
+        XCTAssertEqual(components[3..<5], ["-X", "POST"], "command should contain explicit -X flag")
+        XCTAssertEqual(components.last ?? "", "\"\(URLString)\"", "URL component should be equal")
     }
 
     func testPOSTRequestWithJSONParametersDebugDescription() {
-        let URL = "http://httpbin.org/post"
-        let request = manager.request(.POST, URL, parameters: ["foo": "bar"], encoding: .JSON)
+        // Given
+        let URLString = "http://httpbin.org/post"
+
+        // When
+        let request = manager.request(.POST, URLString, parameters: ["foo": "bar"], encoding: .JSON)
         let components = cURLCommandComponents(request)
 
-        XCTAssert(components[0..<3] == ["$", "curl", "-i"], "components should be equal")
-        XCTAssert(components[3..<5] == ["-X", "POST"], "command should contain explicit -X flag")
-        XCTAssert(request.debugDescription.rangeOfString("-H \"Content-Type: application/json\"") != nil)
-        XCTAssert(request.debugDescription.rangeOfString("-d \"{\\\"foo\\\":\\\"bar\\\"}\"") != nil)
-        XCTAssert(components.last! == "\"\(URL)\"", "URL component should be equal")
+        // Then
+        XCTAssertEqual(components[0..<3], ["$", "curl", "-i"], "components should be equal")
+        XCTAssertEqual(components[3..<5], ["-X", "POST"], "command should contain explicit -X flag")
+        XCTAssertTrue(request.debugDescription.rangeOfString("-H \"Content-Type: application/json\"") != nil, "command should contain 'application/json' Content-Type")
+        XCTAssertTrue(request.debugDescription.rangeOfString("-d \"{\\\"foo\\\":\\\"bar\\\"}\"") != nil, "command data should contain JSON encoded parameters")
+        XCTAssertEqual(components.last ?? "", "\"\(URLString)\"", "URL component should be equal")
     }
 
-    // Temporarily disabled on OS X due to build failure for CocoaPods
-    // See https://github.com/CocoaPods/swift/issues/24
-    #if !os(OSX)
     func testPOSTRequestWithCookieDebugDescription() {
-        let URL = "http://httpbin.org/post"
+        // Given
+        let URLString = "http://httpbin.org/post"
 
         let properties = [
             NSHTTPCookieDomain: "httpbin.org",
@@ -145,17 +265,44 @@ class AlamofireRequestDebugDescriptionTestCase: XCTestCase {
         let cookie = NSHTTPCookie(properties: properties)!
         manager.session.configuration.HTTPCookieStorage?.setCookie(cookie)
 
-        let request = manager.request(.POST, URL)
+        // When
+        let request = manager.request(.POST, URLString)
         let components = cURLCommandComponents(request)
 
-        XCTAssert(components[0..<3] == ["$", "curl", "-i"], "components should be equal")
-        XCTAssert(components[3..<5] == ["-X", "POST"], "command should contain explicit -X flag")
-        XCTAssert(components[5..<6] == ["-b"], "command should contain -b flag")
-        XCTAssert(components.last! == "\"\(URL)\"", "URL component should be equal")
-    }
-    #endif
+        // Then
+        XCTAssertEqual(components[0..<3], ["$", "curl", "-i"], "components should be equal")
+        XCTAssertEqual(components[3..<5], ["-X", "POST"], "command should contain explicit -X flag")
+        XCTAssertEqual(components.last ?? "", "\"\(URLString)\"", "URL component should be equal")
 
-    // MARK: -
+        #if !os(OSX)
+        XCTAssertEqual(components[5..<6], ["-b"], "command should contain -b flag")
+        #endif
+    }
+
+    func testPOSTRequestWithCookiesDisabledDebugDescription() {
+        // Given
+        let URLString = "http://httpbin.org/post"
+
+        let properties = [
+            NSHTTPCookieDomain: "httpbin.org",
+            NSHTTPCookiePath: "/post",
+            NSHTTPCookieName: "foo",
+            NSHTTPCookieValue: "bar",
+        ]
+
+        let cookie = NSHTTPCookie(properties: properties)!
+        managerDisallowingCookies.session.configuration.HTTPCookieStorage?.setCookie(cookie)
+
+        // When
+        let request = managerDisallowingCookies.request(.POST, URLString)
+        let components = cURLCommandComponents(request)
+
+        // Then
+        let cookieComponents = components.filter { $0 == "-b" }
+        XCTAssertTrue(cookieComponents.isEmpty, "command should not contain -b flag")
+    }
+
+    // MARK: Test Helper Methods
 
     private func cURLCommandComponents(request: Request) -> [String] {
         return request.debugDescription.componentsSeparatedByCharactersInSet(NSCharacterSet.whitespaceAndNewlineCharacterSet()).filter { $0 != "" && $0 != "\\" }
