+++ Carthage/Checkouts/Alamofire/Tests/ResponseTests.swift
@@ -1,6 +1,6 @@
 // ResponseTests.swift
 //
-// Copyright (c) 2014–2015 Alamofire (http://alamofire.org)
+// Copyright (c) 2014–2015 Alamofire Software Foundation (http://alamofire.org/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,109 +20,277 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Foundation
 import Alamofire
+import Foundation
 import XCTest
 
-class AlamofireJSONResponseTestCase: XCTestCase {
+class JSONResponseTestCase: BaseTestCase {
     func testGETRequestJSONResponse() {
+        // Given
         let URL = "http://httpbin.org/get"
         let expectation = expectationWithDescription("\(URL)")
 
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var JSON: AnyObject?
+        var error: NSError?
+
+        // When
         Alamofire.request(.GET, URL, parameters: ["foo": "bar"])
-                 .responseJSON { (request, response, JSON, error) in
-                    XCTAssertNotNil(request, "request should not be nil")
-                    XCTAssertNotNil(response, "response should not be nil")
-                    XCTAssertNotNil(JSON, "JSON should not be nil")
-                    XCTAssertNil(error, "error should be nil")
+            .responseJSON { responseRequest, responseResponse, responseJSON, responseError in
+                request = responseRequest
+                response = responseResponse
+                JSON = responseJSON
+                error = responseError
 
-                    XCTAssertEqual(JSON!["args"] as! NSObject, ["foo": "bar"], "args should be equal")
-
-                    expectation.fulfill()
+                expectation.fulfill()
         }
 
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(JSON, "JSON should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        if let args = JSON?["args"] as? NSObject {
+            XCTAssertEqual(args, ["foo": "bar"], "args should match parameters")
+        } else {
+            XCTFail("args should not be nil")
         }
     }
 
     func testPOSTRequestJSONResponse() {
+        // Given
         let URL = "http://httpbin.org/post"
         let expectation = expectationWithDescription("\(URL)")
 
-        Alamofire.request(.POST, URL, parameters: ["foo": "bar"])
-            .responseJSON { (request, response, JSON, error) in
-                XCTAssertNotNil(request, "request should not be nil")
-                XCTAssertNotNil(response, "response should not be nil")
-                XCTAssertNotNil(JSON, "JSON should not be nil")
-                XCTAssertNil(error, "error should be nil")
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var JSON: AnyObject?
+        var error: NSError?
 
-                XCTAssertEqual(JSON!["form"] as! NSObject, ["foo": "bar"], "args should be equal")
+        // When
+        Alamofire.request(.POST, URL, parameters: ["foo": "bar"])
+            .responseJSON { responseRequest, responseResponse, responseJSON, responseError in
+                request = responseRequest
+                response = responseResponse
+                JSON = responseJSON
+                error = responseError
 
                 expectation.fulfill()
         }
 
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(JSON, "JSON should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        if let form = JSON?["form"] as? NSObject {
+            XCTAssertEqual(form, ["foo": "bar"], "form should match parameters")
+        } else {
+            XCTFail("form should not be nil")
         }
     }
 }
 
-class AlamofireRedirectResponseTestCase: XCTestCase {
-    func testGETRequestRedirectResponse() {
-        let URL = "http://google.com"
-        let expectation = expectationWithDescription("\(URL)")
-        
+// MARK: -
+
+class RedirectResponseTestCase: BaseTestCase {
+    func testThatRequestWillPerformHTTPRedirectionByDefault() {
+        // Given
+        let redirectURLString = "http://www.apple.com"
+        let URLString = "http://httpbin.org/redirect-to?url=\(redirectURLString)"
+
+        let expectation = expectationWithDescription("Request should redirect to \(redirectURLString)")
+
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var data: AnyObject?
+        var error: NSError?
+
+        // When
+        Alamofire.request(.GET, URLString)
+            .response { responseRequest, responseResponse, responseData, responseError in
+                request = responseRequest
+                response = responseResponse
+                data = responseData
+                error = responseError
+
+                expectation.fulfill()
+        }
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(data, "data should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        XCTAssertEqual(response?.URL?.URLString ?? "", redirectURLString, "response URL should match the redirect URL")
+        XCTAssertEqual(response?.statusCode ?? -1, 200, "response should have a 200 status code")
+    }
+
+    func testThatRequestWillPerformRedirectionMultipleTimesByDefault() {
+        // Given
+        let redirectURLString = "http://httpbin.org/get"
+        let URLString = "http://httpbin.org/redirect/5"
+
+        let expectation = expectationWithDescription("Request should redirect to \(redirectURLString)")
+
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var data: AnyObject?
+        var error: NSError?
+
+        // When
+        Alamofire.request(.GET, URLString)
+            .response { responseRequest, responseResponse, responseData, responseError in
+                request = responseRequest
+                response = responseResponse
+                data = responseData
+                error = responseError
+
+                expectation.fulfill()
+        }
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(data, "data should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        XCTAssertEqual(response?.URL?.URLString ?? "", redirectURLString, "response URL should match the redirect URL")
+        XCTAssertEqual(response?.statusCode ?? -1, 200, "response should have a 200 status code")
+    }
+
+    func testThatTaskOverrideClosureCanPerformHTTPRedirection() {
+        // Given
+        let redirectURLString = "http://www.apple.com"
+        let URLString = "http://httpbin.org/redirect-to?url=\(redirectURLString)"
+
+        let expectation = expectationWithDescription("Request should redirect to \(redirectURLString)")
         let delegate: Alamofire.Manager.SessionDelegate = Alamofire.Manager.sharedInstance.delegate
-        
-        delegate.taskWillPerformHTTPRedirection = { (session: NSURLSession!, task: NSURLSessionTask!, response: NSHTTPURLResponse!, request: NSURLRequest!) in
-            // Accept the redirect by returning the updated request.
+
+        delegate.taskWillPerformHTTPRedirection = { _, _, _, request in
             return request
         }
-        
-        Alamofire.request(.GET, URL)
-            .response { (request, response, data, error) in
+
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var data: AnyObject?
+        var error: NSError?
+
+        // When
+        Alamofire.request(.GET, URLString)
+            .response { responseRequest, responseResponse, responseData, responseError in
+                request = responseRequest
+                response = responseResponse
+                data = responseData
+                error = responseError
+
                 expectation.fulfill()
-                XCTAssertNotNil(request, "request should not be nil")
-                XCTAssertNotNil(response, "response should not be nil")
-                XCTAssertNotNil(data, "data should not be nil")
-                XCTAssertNil(error, "error should be nil")
-                
-                XCTAssertEqual(response!.URL!, NSURL(string: "http://www.google.com/")!, "request should have followed a redirect")
         }
-        
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
-        }
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(data, "data should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        XCTAssertEqual(response?.URL?.URLString ?? "", redirectURLString, "response URL should match the redirect URL")
+        XCTAssertEqual(response?.statusCode ?? -1, 200, "response should have a 200 status code")
     }
-    
-    func testGETRequestDisallowRedirectResponse() {
-        let URL = "http://google.com/"
-        let expectation = expectationWithDescription("\(URL)")
-        
+
+    func testThatTaskOverrideClosureCanCancelHTTPRedirection() {
+        // Given
+        let redirectURLString = "http://www.apple.com"
+        let URLString = "http://httpbin.org/redirect-to?url=\(redirectURLString)"
+
+        let expectation = expectationWithDescription("Request should not redirect to \(redirectURLString)")
         let delegate: Alamofire.Manager.SessionDelegate = Alamofire.Manager.sharedInstance.delegate
-        
-        delegate.taskWillPerformHTTPRedirection = { (session: NSURLSession!, task: NSURLSessionTask!, response: NSHTTPURLResponse!, request: NSURLRequest!) in
-            // Disallow redirects by returning nil.
-            // TODO: NSURLSessionDelegate's URLSession:task:willPerformHTTPRedirection:newRequest:completionHandler:
-            // suggests that returning nil should refuse the redirect, but this causes a deadlock/timeout
-            
-            return NSURLRequest(URL: NSURL(string: URL)!)
+
+        delegate.taskWillPerformHTTPRedirection = { _, _, _, _ in
+            return nil
         }
-        
-        Alamofire.request(.GET, URL)
-            .response { (request, response, data, error) in
+
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var data: AnyObject?
+        var error: NSError?
+
+        // When
+        Alamofire.request(.GET, URLString)
+            .response { responseRequest, responseResponse, responseData, responseError in
+                request = responseRequest
+                response = responseResponse
+                data = responseData
+                error = responseError
+
                 expectation.fulfill()
-                XCTAssertNotNil(request, "request should not be nil")
-                XCTAssertNotNil(response, "response should not be nil")
-                XCTAssertNotNil(data, "data should not be nil")
-                XCTAssertNil(error, "error should be nil")
-                
-                XCTAssertEqual(response!.URL!, NSURL(string: URL)!, "request should not have followed a redirect")
         }
-        
-        waitForExpectationsWithTimeout(10) { (error) in
-            XCTAssertNil(error, "\(error)")
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(data, "data should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        XCTAssertEqual(response?.URL?.URLString ?? "", URLString, "response URL should match the origin URL")
+        XCTAssertEqual(response?.statusCode ?? -1, 302, "response should have a 302 status code")
+    }
+
+    func testThatTaskOverrideClosureIsCalledMultipleTimesForMultipleHTTPRedirects() {
+        // Given
+        let redirectURLString = "http://httpbin.org/get"
+        let URLString = "http://httpbin.org/redirect/5"
+
+        let expectation = expectationWithDescription("Request should redirect to \(redirectURLString)")
+        let delegate: Alamofire.Manager.SessionDelegate = Alamofire.Manager.sharedInstance.delegate
+        var totalRedirectCount = 0
+
+        delegate.taskWillPerformHTTPRedirection = { _, _, _, request in
+            ++totalRedirectCount
+            return request
         }
+
+        var request: NSURLRequest?
+        var response: NSHTTPURLResponse?
+        var data: AnyObject?
+        var error: NSError?
+
+        // When
+        Alamofire.request(.GET, URLString)
+            .response { responseRequest, responseResponse, responseData, responseError in
+                request = responseRequest
+                response = responseResponse
+                data = responseData
+                error = responseError
+
+                expectation.fulfill()
+        }
+
+        waitForExpectationsWithTimeout(self.defaultTimeout, handler: nil)
+
+        // Then
+        XCTAssertNotNil(request, "request should not be nil")
+        XCTAssertNotNil(response, "response should not be nil")
+        XCTAssertNotNil(data, "data should not be nil")
+        XCTAssertNil(error, "error should be nil")
+
+        XCTAssertEqual(response?.URL?.URLString ?? "", redirectURLString, "response URL should match the redirect URL")
+        XCTAssertEqual(response?.statusCode ?? -1, 200, "response should have a 200 status code")
+        XCTAssertEqual(totalRedirectCount, 5, "total redirect count should be 5")
     }
 }
